Move HTMLDocument to reporting
