+++ nilearn/plotting/__init__.py
@@ -45,7 +45,6 @@ from .html_surface import view_surf, view_img_on_surf
 from .html_stat_map import view_img
 from .html_connectome import view_connectome, view_markers
 from .surf_plotting import plot_surf, plot_surf_stat_map, plot_surf_roi
-from .js_plotting_utils import set_max_img_views_before_warning
 
 __all__ = ['cm', 'plot_img', 'plot_anat', 'plot_epi',
            'plot_roi', 'plot_stat_map', 'plot_glass_brain',
@@ -54,6 +53,5 @@ __all__ = ['cm', 'plot_img', 'plot_anat', 'plot_epi',
            'show', 'plot_matrix', 'view_surf', 'view_img_on_surf',
            'view_img', 'view_connectome', 'view_markers',
            'find_parcellation_cut_coords', 'find_probabilistic_atlas_cut_coords',
-           'plot_surf', 'plot_surf_stat_map', 'plot_surf_roi',
-           'set_max_img_views_before_warning'
+           'plot_surf', 'plot_surf_stat_map', 'plot_surf_roi'
            ]
