+++ nilearn/plotting/html_stat_map.py
@@ -13,10 +13,11 @@ from matplotlib.image import imsave
 from nibabel.affines import apply_affine
 
 from ..image import resample_to_img, new_img_like, reorder_img
-from .js_plotting_utils import get_html_template, HTMLDocument, colorscale
+from .js_plotting_utils import get_html_template, colorscale
 from ..plotting import cm
 from ..plotting.find_cuts import find_xyz_cut_coords
 from ..plotting.img_plotting import _load_anat
+from nilearn.reporting import HTMLDocument
 from .._utils.niimg_conversions import check_niimg_3d
 from .._utils.param_validation import check_threshold
 from .._utils.extmath import fast_abs_percentile
