+++ nilearn/reporting/__init__.py
@@ -2,6 +2,7 @@
 Reporting code for nilearn
 """
 
-from .html_report import ReportMixin
+from .html_report import (ReportMixin, HTMLDocument,
+                          set_max_img_views_before_warning)
 
-__all__ = ['html_report', 'ReportMixin', ]
+__all__ = ['ReportMixin', 'HTMLDocument', 'set_max_img_views_before_warning']
